+++ src/components/templates/club/RosterPage.tsx
@@ -1,5 +1,5 @@
 import Heading from "~/components/atoms/club/Heading";
-import RosterList from "~/components/atoms/club/RosterList";
+import RosterList from "~/components/atoms/club/roster/RosterList";
 import { RosterListWrap, RosterPageLayout } from "./RosterPageStyles";
 import helmet from "~/assets/images/helmet.svg";
 import ROSTER from "~/constants/rosterData";
